feat(Hoare): add a hoare logic for koika
This hoare logic provides a syntax to formalize properties about actions
as hoare triples. These triples consist of a precondition, the action
and a postcondition. To reason over these triple, this framework also
provides hoare lemmas which are build like a weakest precondition
transformer. This way it is possible to programmatically (using an Ltac
tactic) reduce a hoare triple into an implication. For that the
postcondition is transformed into the weakest precondition that needs to
hold (to actually imply the given postcondition). After this reduction
the user only needs to prove that his precondition implies the weakest
precondition.